changed the names of the files for the test to work
